+++ service/persist/contract.go
@@ -7,7 +7,7 @@ import (
 
 // Contract represents an ethereum contract in the database
 type Contract struct {
-	Version      NullInt64       `bson:"version"              json:"version"` // schema version for this model
+	Version      NullInt32       `bson:"version"              json:"version"` // schema version for this model
 	ID           DBID            `bson:"_id"                  json:"id" binding:"required"`
 	CreationTime CreationTime    `bson:"created_at"        json:"created_at"`
 	Deleted      NullBool        `bson:"deleted" json:"-"`
