+++ service/persist/membership.go
@@ -8,7 +8,7 @@ import (
 
 // MembershipTier represents the membership tier of a user
 type MembershipTier struct {
-	Version      NullInt64       `bson:"version"` // schema version for this model
+	Version      NullInt32       `bson:"version"` // schema version for this model
 	ID           DBID            `bson:"_id"           json:"id" binding:"required"`
 	CreationTime CreationTime    `bson:"created_at" json:"created_at"`
 	Deleted      NullBool        `bson:"deleted" json:"-"`
