+++ service/persist/nft.go
@@ -9,7 +9,7 @@ import (
 
 // NFT represents an nft throughout the application
 type NFT struct {
-	Version         NullInt64       `bson:"version"              json:"version"` // schema version for this model
+	Version         NullInt32       `bson:"version"              json:"version"` // schema version for this model
 	ID              DBID            `bson:"_id"                  json:"id" binding:"required"`
 	CreationTime    CreationTime    `bson:"created_at"        json:"created_at"`
 	Deleted         NullBool        `bson:"deleted" json:"-"`
