+++ service/persist/token.go
@@ -142,7 +142,7 @@ type AddressAtBlock struct {
 
 // Token represents an individual Token token
 type Token struct {
-	Version      NullInt64       `bson:"version"              json:"version"` // schema version for this model
+	Version      NullInt32       `bson:"version"              json:"version"` // schema version for this model
 	ID           DBID            `bson:"_id"                  json:"id" binding:"required"`
 	CreationTime CreationTime    `bson:"created_at"        json:"created_at"`
 	Deleted      NullBool        `bson:"deleted" json:"-"`
